index 위젯 분리
