+++ app/(tabs)/index.tsx
@@ -1,19 +1,12 @@
-import DateTimePicker, {
-  DateTimePickerEvent,
-} from "@react-native-community/datetimepicker";
+import { InformationBox } from "@/components/home/InformationBox";
+import { MapWidget } from "@/components/home/MapWidget";
+import { 방_탐색 } from "@/components/home/방_탐색";
+import { 방장_하기 } from "@/components/home/방장_하기";
+import { DateTimePickerEvent } from "@react-native-community/datetimepicker";
 import * as Location from "expo-location";
-import { Link } from "expo-router";
 import React, { useEffect, useRef, useState } from "react";
-import {
-  Dimensions,
-  FlatList,
-  Modal,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
-} from "react-native";
-import MapView, { Marker, Polyline } from "react-native-maps";
+import { Dimensions, StyleSheet, View } from "react-native";
+import MapView from "react-native-maps";
 
 const { width } = Dimensions.get("window"); //Dimensions 이용
 
@@ -152,210 +145,48 @@ export default function HomeView() {
 
   return (
     <View style={styles.container}>
-      {/* 지도 상자 */}
-      <View style={map_styles.mapInfoBox}>
-        {location ? (
-          <MapView
-            ref={mapRef} // 지도 참조 추가
-            style={map_styles.map}
-            initialRegion={{
-              latitude: location.latitude,
-              longitude: location.longitude,
-              latitudeDelta: 0.05,
-              longitudeDelta: 0.05,
-            }}
-          >
-            <Marker
-              coordinate={{
-                latitude: coordinates[selectedDeparture].latitude,
-                longitude: coordinates[selectedDeparture].longitude,
-              }}
-              title="출발지"
-            />
-            <Marker
-              coordinate={{
-                latitude: coordinates[selectedDestination].latitude,
-                longitude: coordinates[selectedDestination].longitude,
-              }}
-              title="도착지"
-            />
-            {route && (
-              <Polyline
-                coordinates={route}
-                strokeColor="#6B59CC"
-                strokeWidth={4}
-              />
-            )}
-          </MapView>
-        ) : (
-          <View style={map_styles.loadingContainer}>
-            <Text>위치를 불러오는 중...</Text>
-            {errorMsg ? <Text>{errorMsg}</Text> : null}
-          </View>
-        )}
-      </View>
+      <MapWidget
+        location={location}
+        coordinates={coordinates}
+        selectedDeparture={selectedDeparture}
+        selectedDestination={selectedDestination}
+        route={route}
+        errorMsg={errorMsg}
+        mapRef={mapRef}
+      />
 
-      {/* 상호작용 UI */}
       <View style={styles.interface}>
-        <View style={main_styles.infoBox}>
-          {/* 첫번째 열 */}
-          <View
-            style={{
-              flexDirection: "row",
-            }}
-          >
-            {/* 출발지 */}
-            <View style={{ flex: 3 }}>
-              <Text style={main_styles.locationLabel}>출발</Text>
-              <TouchableOpacity onPress={() => openLocationModal("departure")}>
-                <Text style={main_styles.routeText}>{selectedDeparture}</Text>
-              </TouchableOpacity>
-            </View>
-
-            {/* 화살표 */}
-            <View style={{ flex: 1 }}>
-              {/* 빈 텍스트로 공간 확보 */}
-              <Text></Text>
-              <Text style={main_styles.arrow}>➜</Text>
-            </View>
-
-            {/* 도착지 */}
-            <View style={{ flex: 3 }}>
-              <Text style={main_styles.locationLabel}>도착</Text>
-              <TouchableOpacity
-                onPress={() => openLocationModal("destination")}
-              >
-                <Text style={main_styles.routeText}>{selectedDestination}</Text>
-              </TouchableOpacity>
-            </View>
-
-            {/* 출발지와 목적지에 관한 모달 */}
-            <Modal
-              animationType="slide"
-              transparent={true}
-              visible={modalVisible}
-              onRequestClose={() => setModalVisible(false)}
-            >
-              <View style={choice_styles.modalView}>
-                <FlatList
-                  data={locations.filter(
-                    (item) =>
-                      (changingLocationType === "departure" &&
-                        item !== selectedDestination) ||
-                      (changingLocationType === "destination" &&
-                        item !== selectedDeparture),
-                  )}
-                  keyExtractor={(item) => item}
-                  renderItem={({ item }) => (
-                    <TouchableOpacity
-                      style={choice_styles.item}
-                      onPress={() => handleLocationSelect(item)}
-                    >
-                      <Text style={choice_styles.itemText}>{item}</Text>
-                    </TouchableOpacity>
-                  )}
-                  ItemSeparatorComponent={() => (
-                    <View style={choice_styles.separator} />
-                  )}
-                />
-              </View>
-            </Modal>
-          </View>
-
-          {/* 두번째 열 */}
-          <View>
-            {/* 출발 시간 */}
-            <Text style={date_styles.infoTitle}>출발 시간</Text>
-            <View style={date_styles.dateTimeRow}>
-              <TouchableOpacity
-                onPress={() => setShowDatePicker(true)}
-                style={date_styles.dateTimeButton}
-              >
-                <Text style={date_styles.dateText}>
-                  {date.toLocaleDateString()}
-                </Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                onPress={() => setShowTimePicker(true)}
-                style={date_styles.dateTimeButton}
-              >
-                <Text style={date_styles.dateText}>
-                  {date.toLocaleTimeString([], {
-                    hour: "2-digit",
-                    minute: "2-digit",
-                  })}
-                </Text>
-              </TouchableOpacity>
-            </View>
-          </View>
-
-          {showDatePicker && (
-            <DateTimePicker
-              value={date}
-              mode="date"
-              display="default"
-              onChange={handleDateChange}
-            />
-          )}
-
-          {showTimePicker && (
-            <DateTimePicker
-              value={date}
-              mode="time"
-              display="default"
-              minuteInterval={5}
-              onChange={handleTimeChange}
-            />
-          )}
-
-          {/* 세번째 열 */}
-          {/* 경로 정보 표시 */}
-          {distance && duration && (
-            <View
-              style={{
-                flexDirection: "row",
-                padding: 10,
-                backgroundColor: "#F1E7FC",
-                gap: 10,
-              }}
-            >
-              <Text>거리: {distance.toFixed(2)} km</Text>
-              <Text>소요 시간: {duration.toFixed(2)} 분</Text>
-            </View>
-          )}
-        </View>
+        {/* 상호작용 UI */}
+        <InformationBox
+          selectedDeparture={selectedDeparture}
+          selectedDestination={selectedDestination}
+          date={date}
+          showDatePicker={showDatePicker}
+          showTimePicker={showTimePicker}
+          modalVisible={modalVisible}
+          changingLocationType={changingLocationType}
+          locations={locations}
+          distance={distance}
+          duration={duration}
+          openLocationModal={openLocationModal}
+          setModalVisible={setModalVisible}
+          handleDateChange={handleDateChange}
+          handleTimeChange={handleTimeChange}
+          handleLocationSelect={handleLocationSelect}
+        />
 
         {/* 버튼 */}
         <View style={styles.buttonContainer}>
-          {/* 방 탐색 버튼 */}
-          <Link
-            href={{
-              pathname: "/RoomList",
-              params: {
-                selectedDeparture,
-                selectedDestination,
-                date: date.toISOString(),
-              },
-            }}
-            style={styles.button}
-          >
-            <Text style={styles.buttonText}>방 탐색</Text>
-          </Link>
-
-          {/* 방장 하기 버튼 */}
-          <Link
-            href={{
-              pathname: "/RoomMake",
-              params: {
-                selectedDeparture,
-                selectedDestination,
-                date: date.toISOString(),
-              },
-            }}
-            style={styles.button}
-          >
-            <Text style={styles.buttonText}>방장 하기</Text>
-          </Link>
+          <방_탐색
+            departure={selectedDeparture}
+            destination={selectedDestination}
+            date={date}
+          />
+          <방장_하기
+            departure={selectedDeparture}
+            destination={selectedDestination}
+            date={date}
+          />
         </View>
       </View>
     </View>
@@ -389,130 +220,5 @@ const styles = StyleSheet.create({
   buttonText: {
     color: "#FFFFFF",
     fontSize: width * 0.045,
-  },
-});
-
-const map_styles = StyleSheet.create({
-  // 지도화면, 소요시간 글씨
-  mapInfoBox: {
-    // 절대 위치로 지정해서 지도가 배경으로 나타나게 함
-    position: "absolute",
-    top: 0,
-    left: 0,
-    right: 0,
-    height: "75%",
-    zIndex: 0,
-    borderWidth: 1,
-    borderColor: "#ccc",
-  },
-  map: {
-    ...StyleSheet.absoluteFillObject,
-  },
-  loadingContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-});
-
-const main_styles = StyleSheet.create({
-  //출발지 도착지
-  infoBox: {
-    backgroundColor: "#FFFFFF",
-    padding: 10,
-    borderRadius: 8,
-    borderWidth: 1,
-    borderColor: "#ccc",
-    shadowColor: "#000",
-    shadowOpacity: 0.1,
-    shadowRadius: 6,
-    elevation: 3,
-    gap: 10,
-  },
-  locationSection: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    width: "70%",
-    marginBottom: 10,
-  },
-  locationLabel: {
-    textAlign: "center",
-    fontSize: width * 0.037,
-    color: "#888",
-  },
-
-  locationSelector: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-evenly",
-    marginBottom: "8%",
-    width: "120%",
-  },
-  routeContainer: {
-    width: width * 0.3, // 고정된 너비로 박스가 항상 일정하게 유지
-    alignItems: "center",
-  },
-  routeText: {
-    textAlign: "center",
-    fontSize: width * 0.06,
-    color: "#6B59CC",
-    fontWeight: "bold",
-  },
-  arrow: {
-    textAlign: "center",
-    textAlignVertical: "center",
-    fontSize: width * 0.06,
-    fontWeight: "bold",
-    color: "#6B59CC",
-  },
-});
-
-const choice_styles = StyleSheet.create({
-  //출발지 도착지 선택 항목
-  modalView: {
-    flex: 1,
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-    paddingVertical: "10%",
-    paddingHorizontal: "10%",
-    alignItems: "stretch",
-  },
-  item: {
-    backgroundColor: "white",
-    padding: "5%",
-    marginVertical: "1.5%",
-    borderWidth: 1,
-    borderColor: "#ccc",
-    borderRadius: 10,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  separator: {
-    height: 1,
-  },
-});
-
-const date_styles = StyleSheet.create({
-  // 날짜, 시간
-  infoTitle: {
-    fontSize: width * 0.037,
-    fontWeight: "bold",
-    color: "#888",
-  },
-  dateTimeRow: {
-    flexDirection: "row",
-    gap: 20,
-  },
-  dateTimeButton: {
-    backgroundColor: "#FFFFFF",
-    borderRadius: 5,
-  },
-  dateText: {
-    fontSize: width * 0.06,
-    color: "#6B59CC",
-  },
-  loadingContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
   },
 });
