+++ components/home/button_style.tsx
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  button: {
+    flex: 1,
+    backgroundColor: "#7963F4",
+    borderRadius: 8,
+    textAlign: "center",
+    textAlignVertical: "center",
+  },
+  text: {
+    color: "#FFFFFF",
+    fontSize: 18,
+    fontWeight: "bold",
+  },
+});
